Update config.md
It looks like https://github.com/FormidableLabs/react-native-app-auth/issues/1003, the issue was resolved by adding a trailing slash, and I may have seen another issue or two regarding this elsewhere. In fact, we already call it out in the specific provider section, but this might help surface it a bit more too.